requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
cymem==2.0.5
feedparser==6.0.8
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.21.1
packaging==21.0
pathy==0.6.0
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
SQLAlchemy==1.4.22
srsly==2.4.1
thinc==8.0.8
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
